agent/pyproject.toml
[project]
name = "bro-agent"
version = "0.1.0"
description = "LiveKit STT Agent for Bro Voice Recorder"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=1.0.0",
    "livekit-plugins-silero",
    "livekit-plugins-deepgram",
    "livekit-plugins-elevenlabs",
    "python-dotenv",
]
